backend/requirements.txt
langchain==0.3.7
langchain-community==0.3.5
langchain-groq==0.2.1
langchain-huggingface==0.1.2
faiss-cpu==1.9.0
sentence-transformers==3.3.1
pypdf==5.1.0
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-dotenv==1.0.1
pydantic==2.10.2
requests==2.32.3
redis==5.2.0
reportlab==4.2.5
